Initial Commit for BooksApi Final Commit for Contoso University

+++ ContosoUniversity/ContosoUniversity/Controllers/CoursesController.cs
@@ -29,7 +29,7 @@ namespace ContosoUniversity.Controllers
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
-        public async Task<IActionResult> UpdateCourseCredits(int? multiplier)
+        public async Task<IActionResult> UpdateCourseCredits(float? multiplier)
         {
             if (multiplier != null)
             {
